fix(dia.LinkView): Paper sometimes autosizes incorrectly after moving links with doubleLinkTools option
Moving around elements and links that have doubleLinkTools: true but link length is short enough
for the secondary tool to be hidden, the tool position is not updated, causing miscalculations in paper size and paddings.

Fix by using display instead of visibility, to take the tool out of the rendering tree when invisible.