fix(dia.Paper): don't leak animation frame request
When `resetViews()` calls `_resetUpdates()` it's possible that there's an animation frame request, but `resetUpdates()` was not checking for this and cancelling any such request, potentially causing `updateViewsAsync()` to be called when it shouldn't be.

This can cause unexpected behavior if the paper was frozen, or cause an error to be thrown if the paper was destroyed.